Mac kext: Hides PrjFSProviderUserClient API behind free functions for mocking
The PrjFSProviderUserClient class, being an OSObject based class is near-impossible to use properly from user space. This moves all uses of it from the rest of the kext behind free functions which we can easily replace with mocked versions when unit testing. VirtualizationRoots.cpp then only needs a forward declaration of the class.